requirements.txt
discord.py
selenium
webdriver-manager
beautifulsoup4
requests-html
lxml
matplotlib
sqlite3